fix: Keep Footshop off /ico and render its OG card unhovered
/ico is about the sole proprietorship, so it shouldn't carry the employer:
the site-wide Person structured data (with `worksFor: Footshop`) moves out of
the root layout into a PersonJsonLd component rendered by the home, CV and
project routes, and /ico declares its own keywords instead of inheriting the
portfolio set.

The OG render is now flat and unhovered — screenshotted with reduced motion
emulated, which parks the tilt ticker and the drift animations.
