+++ src/app/cv/page.tsx
@@ -1,8 +1,9 @@
-import {CV} from '@/components'
+import {CV, PersonJsonLd} from '@/components'
 
 export default async function Cv() {
   return (
     <div className="flex w-full h-full justify-center overflow-y-scroll pt-[116px]">
+      <PersonJsonLd />
       <div className="max-w-[572px] w-full h-full flex flex-col items-center">
         <CV animated>
           <div className="flex h-[116px] w-full flex-shrink-0" />
