+++ src/components/index.ts
@@ -11,6 +11,7 @@ export * from './cv'
 export * from './google-analytics'
 export * from './ico-card'
 export * from './intro-overlay'
+export * from './person-json-ld'
 export * from './shot-detail'
 export * from './sound-controls'
 export * from './web-vitals'
