+++ src/components/person-json-ld.tsx
@@ -1,0 +1,39 @@
+const jsonLd = {
+  '@context': 'https://schema.org',
+  '@type': 'Person',
+  address: {
+    '@type': 'PostalAddress',
+    addressCountry: 'CZ',
+    addressLocality: 'Prague',
+  },
+  email: 'hello@janblazej.dev',
+  image: 'https://janblazej.dev/png/profile@3x.png',
+  jobTitle: 'Lead Mobile Developer & Founder',
+  knowsAbout: [
+    'React Native',
+    'Expo',
+    'TypeScript',
+    'iOS Development',
+    'Android Development',
+    'Mobile App Development',
+    'Swift',
+    'visionOS',
+    'RealityKit',
+  ],
+  name: 'Jan Blazej',
+  sameAs: ['https://www.linkedin.com/in/hzblj', 'https://github.com/hzblj', 'https://www.instagram.com/hzblj'],
+  url: 'https://janblazej.dev',
+  worksFor: {
+    '@type': 'Organization',
+    name: 'Footshop',
+    url: 'https://footshop.com',
+  },
+}
+
+/**
+ * Personal identity structured data. Rendered per route rather than site-wide,
+ * so pages about the sole proprietorship (/ico) aren't tied to an employer.
+ */
+export const PersonJsonLd = () => (
+  <script type="application/ld+json" dangerouslySetInnerHTML={{__html: JSON.stringify(jsonLd)}} />
+)
